move dirs into ./client
