+++ client/index.jsx
@@ -4,8 +4,8 @@ import { createStore } from 'redux'
 import ReactDOM from "react-dom"
 import React from "react"
 
-import App from '../containers/App'
-import configure from '../store'
+import App from './containers/App'
+import configure from './store'
 
 const store = configure()
